A very basic CSS animation
Hey! This is my first CSS animation and it is created for the web dev course I am currently learning. This animation basically have a circle moving around the screen borders infinitely, with some text placed onto the centre of the circle. Also note that this animation is not responsive and might not work well for different-sized screens; where it's best to display on 13-inch laptops.